trace-cruncher: Add initial Dockerfile
This commit continues work towards making trace-cruncher fully reproducible and containerized.

The Dockerfile builds a Debian image with trace-cruncher installed, using the new date-snapshot script to freeze the built dependencies.

The next phase of this work will be moving to a Debuerreotype base image so that the container build will be completely reproducible.

Currently the Docker image pulls from the knauth/trace-cruncher fork. This will be changed to the main VMware repository once the date-snapshot script is merged.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>